requirements.txt
astroid==2.4.2
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
colorthief==0.2.1
facebook-sdk==3.1.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.1.2
pylint==2.5.3
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
requests==2.24.0